GPS and all I2C sensors working
GPS is now working, UART is completely functional and an example is given how to transmit the GPS data. The configuration is also sent to the GPS to only get the one line of data we wanted.

I2C sensors are now all working and saved in variables. Code was commented for readability.